requirements.txt
kafka-python==2.0.2
requests==2.25.1
psycopg2-binary==2.8.6
pytest==6.2.2
pytest-html-reporter==0.2.3
apscheduler==3.7.0